Add privilage to edit and delete only for authors of posts

+++ base/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
 from django.contrib import messages
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic
@@ -31,8 +33,7 @@ def loginUser(request):
         else:
             messages.error(request, "User doesn't exist")
 
-    context = {}
-    return render(request, 'base/login_register.html', context)
+    return render(request, 'base/login_register.html')
 
 def logoutUser(request):
     logout(request)
@@ -58,6 +59,7 @@ def room(request, pk):
     context = {'room':room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
 
@@ -70,10 +72,14 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
     
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid:
@@ -83,8 +89,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
@@ -93,7 +104,6 @@ def deleteRoom(request, pk):
     return render(request, 'base/delete.html', context)
 
 def addUser(request):
-
     form = UserForm()
     if request.method == 'POST':
         form = UserForm(request.POST)
